+++ src/main/java/tk/hiddenname/smarthome/controller/sensor/SensorSignalRestController.java
@@ -5,8 +5,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import tk.hiddenname.smarthome.entity.hardware.Sensor;
-import tk.hiddenname.smarthome.entity.signal.DigitalState;
+import tk.hiddenname.smarthome.model.hardware.Sensor;
+import tk.hiddenname.smarthome.model.signal.DigitalState;
 import tk.hiddenname.smarthome.service.database.SensorDatabaseService;
 import tk.hiddenname.smarthome.service.hardware.impl.digital.input.DigitalSensorService;
 
@@ -24,7 +24,7 @@ public class SensorSignalRestController {
     // ***************************** DIGITAL **************************************************
 
     @GetMapping(value = {"/digital"}, produces = {"application/json"})
-    public DigitalState getState(@Valid @RequestParam(name = "id") Integer id) {
+    public DigitalState getState(@Valid @RequestParam(name = "id") Long id) {
         Sensor sensor = dbService.getOne(id);
 
         return digitalService.getState(sensor.getId(), sensor.isReverse());
