+++ src/main/java/tk/hiddenname/smarthome/exception/SensorNotFoundException.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
 public class SensorNotFoundException extends RuntimeException {
 
-    public SensorNotFoundException(int id) {
+    public SensorNotFoundException(Long id) {
         super(String.format("The sensor with id '%d' not found", id));
     }
 }
