+++ src/main/java/tk/hiddenname/smarthome/service/hardware/impl/digital/output/DigitalDeviceService.java
@@ -1,11 +1,11 @@
 package tk.hiddenname.smarthome.service.hardware.impl.digital.output;
 
-import tk.hiddenname.smarthome.entity.signal.DigitalState;
+import tk.hiddenname.smarthome.model.signal.DigitalState;
 import tk.hiddenname.smarthome.service.hardware.impl.GPIOService;
 
 public interface DigitalDeviceService extends GPIOService {
 
-    DigitalState getState(Integer id, Boolean reverse);
+    DigitalState getState(Long id, Boolean reverse);
 
-    DigitalState setState(Integer id, Boolean reverse, Boolean newState);
+    DigitalState setState(Long id, Boolean reverse, Boolean newState);
 }
